+++ src/main/kotlin/app/revanced/patches/youtube/layout/buttons/player/hide/patch/HidePlayerButtonsPatch.kt
@@ -31,6 +31,7 @@ class HidePlayerButtonsPatch : BytecodePatch(
 ) {
     private object ParameterOffsets {
         const val HAS_NEXT = 5
+        const val HAS_PREVIOUS = 6
     }
 
     override fun execute(context: BytecodeContext): PatchResult {
@@ -58,13 +59,16 @@ class HidePlayerButtonsPatch : BytecodePatch(
 
             // overriding this parameter register hides the previous and next buttons
             val hasNextParameterRegister = callInstruction.startRegister + ParameterOffsets.HAS_NEXT
+            val hasPreviousParameterRegister = callInstruction.startRegister + ParameterOffsets.HAS_PREVIOUS
 
             mutableMethod.addInstructions(
                 callIndex,
                 """
-                    invoke-static { }, Lapp/revanced/integrations/patches/HidePlayerButtonsPatch;->hideButtons()Z
+                    invoke-static { v$hasNextParameterRegister }, Lapp/revanced/integrations/patches/HidePlayerButtonsPatch;->previousOrNextButtonIsVisible(Z)Z
                     move-result v$hasNextParameterRegister
-                    xor-int/lit8 v$hasNextParameterRegister, v$hasNextParameterRegister, 1
+                    
+                    invoke-static { v$hasPreviousParameterRegister }, Lapp/revanced/integrations/patches/HidePlayerButtonsPatch;->previousOrNextButtonIsVisible(Z)Z
+                    move-result v$hasPreviousParameterRegister
                 """
             )
         } ?: return PlayerControlsVisibilityModelFingerprint.toErrorResult()