+++ src/main/kotlin/app/revanced/patches/youtube/misc/integrations/fingerprints/RemoteEmbeddedPlayerFingerprint.kt
@@ -1,0 +1,19 @@
+package app.revanced.patches.youtube.misc.integrations.fingerprints
+
+import app.revanced.patcher.extensions.or
+import app.revanced.patches.shared.integrations.patch.AbstractIntegrationsPatch.IntegrationsFingerprint
+import org.jf.dexlib2.AccessFlags
+
+/**
+ * For embedded playback inside 3rd party android app (such as 3rd party Reddit apps).
+ */
+object RemoteEmbeddedPlayerFingerprint : IntegrationsFingerprint(
+    accessFlags = AccessFlags.PRIVATE or AccessFlags.CONSTRUCTOR,
+    returnType = "V",
+    parameters = listOf("Landroid/content/Context;", "L", "L", "Z"),
+    customFingerprint = { methodDef, _ ->
+        methodDef.definingClass == "Lcom/google/android/youtube/api/jar/client/RemoteEmbeddedPlayer;"
+    },
+    // Integrations context is the first method parameter.
+    contextRegisterResolver = { it.implementation!!.registerCount - it.parameters.size }
+)